[MOV-038] Anamneses e aplicaçao ok
